refactor: overlay components into plugin format
