+++ apps/page-builder-demo/src/components/overlay-plugins/image-res.tsx
@@ -1,5 +1,6 @@
 import {get} from '@sanity/util/paths'
 import {OverlayPluginDefinition, useDocuments} from '@sanity/visual-editing/react'
+import {defineOverlayPlugin} from '@sanity/visual-editing/unstable_overlay-components'
 import {useEffect, useState} from 'react'
 
 export function parseImageAssetId(documentId: string) {
@@ -9,7 +10,7 @@ export function parseImageAssetId(documentId: string) {
   return {type: 'image', assetId, width, height, extension}
 }
 
-export const ImageResolutionHUD: OverlayPluginDefinition = {
+export const ImageResolutionHUD = defineOverlayPlugin((options) => ({
   type: 'hud',
   name: 'image-resolution-hud',
   component: function ImageResolutionHUDComponent(props) {
@@ -41,4 +42,4 @@ export const ImageResolutionHUD: OverlayPluginDefinition = {
     // @ts-ignore
     return ctx?.field?.value?.fields?.asset?.value?.dereferencesTo === 'sanity.imageAsset'
   },
-}
+}))
