+++ apps/page-builder-demo/src/components/overlay-plugins/index.tsx
@@ -1,14 +1,20 @@
 'use client'
 
 import {OverlayPluginDefinition} from '@sanity/visual-editing/react'
-import {ExampleExclusivePlugin} from './example-exclusive'
-import {ExampleHUDPlugin} from './example-hud'
+import {ExcitingTitlePlugin} from './exciting-title'
 import {ImageResolutionHUD} from './image-res'
 import {LEDLifespanHUD} from './led-lifespan'
+import {Rotate3D} from './rotate-3d'
 
 export const plugins: OverlayPluginDefinition[] = [
-  LEDLifespanHUD,
-  ImageResolutionHUD,
-  ExampleExclusivePlugin,
-  ExampleHUDPlugin,
+  LEDLifespanHUD(),
+  ImageResolutionHUD(),
+  Rotate3D({guard: ({node}) => node.path.endsWith('rotations')}),
+  ExcitingTitlePlugin({
+    guard: ({node}) => node.path.endsWith('title'),
+    options: {append: '?', buttonText: '🤩'},
+  }),
+  ExcitingTitlePlugin({
+    guard: ({node}) => node.path.endsWith('headline'),
+  }),
 ]
